requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
pymongo==4.0.2
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.1.1
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
